ruff_python_parser: remove unchecked conversion to str
While the usage looks correct, the use of `unsafe` here does not seem
justified to me. Namely, it's already doing integer parsing. And perhaps
most importantly, this is for parsing an octal literal which are likely
to be rare enough to not have a major impact on perf. (And it's not like
UTF-8 validation is slow.)

This was originally introduced in #8227 and it doesn't look like
unchecked string conversion was the main point there.