ruff_python_formatter: tweak AST normalization
This changes the AST normalizer to remove anything that looks like a
code snippet. The idea here is that, at least in tests, we want to check
that reformatting Python code does *not* change its AST. Well, it is
supposed to change its AST in some ways, and this normalizer tries to
erase those changes.

It turns out that reformatting code examples in docstrings will change
the AST in even more profound ways. Namely, it can arbitrarily rewrite
docstring contents. Because of this, it doesn't really seem feasible to
normalize the strings in any way other than to remove anything that
looks like a code snippet.