ruff_python_formatter: add docstring state to context
This augments the Python's formatter context to include
state about a docstring's quote style. Namely, the quote
style refers to the kind of quotes used for a docstring
that contains a code example that is currently being
formatted. This state will allow us to choose the correct
quote style to use while reformatting code snippets and
will avoid writing invalid Python (in most cases).