ruff_python_formatting: add doctest state handling
This commit adds a small but central component to code snippet
formatting in docstrings: it specifically implements the state
transitions needed to recognize and collect code snippets from
doctests. This means looking for PS1 and PS2 prompts and extracting
the code portion of each line.

This also introduces a "code example add action" which we will
use in a subsequent commit to control the higher level docstring
line printer.