ruff_python_formatter: add 'line-ending' to test output
It looks like no previous snapshot tests have configured the line ending
to anything other than the default, so this wasn't included in the test
output. But we would actually like to try and test that line endings are
correctly preserved when reformatting code snippets, so we make the
option visible in the snapshot.