Add `Invalid` node for `Expr` and `Pattern`
These nodes are going to be used by the new parser to represent invalid syntax.

These nodes are handled in the `ruff_python_formatter` by emitting the text as
it is written in the source code.

For the type inference in the `ruff_python_semantic`, the new nodes are resolved
to `Unknown`.

In the `ruff_python_codegen` they are emitted as it is written in the source code.