ruff_python_formatter: slightly refactor doctest formatting
This moves some free functions to methods on the CodeExampleDoctest
type. This bundles some of the operations for finding prompt lines with
the state required.

I was motivated to do this by in-progress work on supporting reStructuredText
blocks. It wants a bit more state, and buttoning things up into methods
helps make the code clearer I think.

Note that this PR introduces a few mor to_string() and clone() calls.
I'm going to try to remove those in a subequent commit.