ruff_python_formatter: split DocstringLine in two
This refactors the DocstringLine type into input and output types. This
lets us elide the Cow from the input docstring line, and that in turn
lets us borrow that input line in more places. For example, as done in
this commit, it lets us change the `code` field of a `CodeExampleLine`
to a borrowed string. (I had previously made it a `String` in part
because of borrow issues, but also because I thought there might be
transformations made to it. But after further thought, I think a code
line should always be a substring of the original line
pre-reformatting.)

This refactor is motivated by wanting to add some more state to the
"input" docstring line. For example, the reStructuredText handling wants
to know what the next line is in order to terminate a block. Adding that
to `InputDocstringLine` is no problem, but it doesn't make sense on the
`OutputDocstringLine`---especially as a borrowed string---since the
output line may be the result of reformatting.

Finally, this lets us make `InputDocstringLine` impl `Copy` and avoids
some clones and to_string() calls. (I don't think it's likely to make a
perf difference, but it does seem cleaner.)