ruff_python_formatter: switch 'is_last' to 'next'
That is, for the input docstring line, instead of just recording whether
a particular line is the last one or not, we should just include the
next line. If one doesn't exist, then we know it's the last line.

This will be useful for recognizing code snippets when knowing the next
line is required for determining the something about the code snippet
(such as where a reStructuredText literal block ends).