ruff_python_formatter: refactor where code lines are stored
Basically, instead of putting lines on CodeExample, we put them on the
individual CodeExampleKind. The reason for this is that the code lines
ought to be owned by the specific kind of code example that is being
parsed.

For example, with reStructuredText, the actual content of each code line
cannot be known until the entire block has been parsed. This is because
the entire block needs to be seen before the correct minimal indent
level is known.

By putting the code lines into each individual kind, we allow the kind
to "own" the logic for dealing with code lines.