ruff_python_formatter: fix bug in doctest formatting
This fixes a bug where if the very last line in a docstring is a
doctest, then it wasn't getting formatted.

We do a little more than fix that bug in this commit. In particular, we
industrialize how we handle actions by using a queue. Instead of trying
to do too much with each action, we make each one a bit simpler and
build the infrastructure necessary to permit more arbitrary composition.
This in particular lets us handle weird corner cases like "the code
example and the docstring end at the same time" without much fuss.

In essence, actions like Format and Reset no longer carry the "original"
line. Instead of doing that, we generate two actions: the first to
format and the second (if necessary) to print the original line.

This infrastructure will become more apparently useful when dealing with
reStructuredText blocks.