Replace autogenerated parser with hand-written parser
This commit replaces the autogenerated lalrpop parser in favor of a handwritten
parser. Beyond the performance improvements, one of the main differences between
the handwritten parser and the autogenerated parser is, the new parser is error
persistent. That is, the parser can create an AST even if the source code contains
invalid syntax. This also means that, the parser now collects all syntax errors
found in the source code, while the old parser would stop parsing after finding
the first invalid syntax.

The new parser still preserves the old behavior of returning only the first syntax
error it encounters. The Python syntax supported by the parser is of the version 3.12.

There is a 14.33% (if my math is correct) overall performance improvement with the new parser.

Refactor `ParseError` to use `TextRange` for the error location instead of `TextSize`.

The following `LexicalErrorType` were moved to `ParseErrorType`:
  - `PositionalArgumentError`
  - `UnpackedArgumentError`
  - `DuplicateArgumentError`
  - `DuplicateKeywordArgumentError`
  - `AssignmentError`
These errors arise during the parsing phase rather than the lexing phase.

The `StringType` is a helper type used when parsing string/f-string literals. Since the
parser now handles invalid syntax, we need a way to represent an invalid string/f-string
when parsing string/f-string literals. So we add the `Invalid` type to deal with that.

Here's how the parser handles invalid strings:

For a single invalid string literal, e.g., a string with invalid escape sequence `"a \xxx"`,
the resulting AST node for that will be an `Expr::Invalid`. The current implementation for
string literal parsing has one limitation when it comes to implicit concatenated strings.

We can't know yet, which specific string within the implicit concatenated strings are invalid.
That is, the resulting AST node won't have an `Invalid` node for the invalid string literal.

Additionally, the parser now parses the syntax mentioned in issue #8914.