ruff_python_formatter: add new docstring-code-line-width option
This permits setting a line width for code examples in docstrings
that is distinct from the line width of the surrounding Python
code. Its default is the same as for Python: 88. But note that
changing the line width for the surrounding Python code does not
also change it for docstring code examples. They are configured
separately.

Note that I chose "width" for this internal option to make it
consistent with other internal options, even though it seems like
we would like to eventually use the word "length" to match public
facing documentation.