ruff_python_formatter: add a new docstring test configuration
We add a new configuration with a docstring line width, but specifically
keep it at the default for now to make review changes based on line
width easier.

In the next commit, will add changes to the snapshot.

In the commit following that, we'll finally change the actual line width
setting and we'll be able to see precisely its impact.